+++ crates/cast/src/helpers/constants.rs
@@ -24,3 +24,4 @@ pub const DEFAULT_RETRIES: u8 = 12;
 pub const DEFAULT_ACCOUNTS_FILE: &str = "~/.starknet_accounts/starknet_open_zeppelin_accounts.json";
 
 pub const KEYSTORE_PASSWORD_ENV_VAR: &str = "KEYSTORE_PASSWORD";
+pub const TEST_KEYSTORE_PASSWORD_ENV_VAR: &str = "TEST_KEYSTORE_PASSWORD";