+++ CHANGELOG.md
@@ -22,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 - `account add` command for importing accounts to the accounts file
-- `account create --keystore` command for creating openzeppelin accounts with keystore
+- `account create` command for creating openzeppelin accounts with starkli-style keystore
 
 ### Changed